backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
yfinance
python-dotenv
httpx
